+++ node.d.ts
@@ -1,0 +1,6 @@
+import { NodeLogger } from './src';
+
+declare const Logger: NodeLogger & {
+  default: NodeLogger;
+};
+export = Logger;