+++ node.js
@@ -1,0 +1,5 @@
+'use strict';
+
+const node = require('./src/node');
+
+module.exports = node;