+++ src/node/index.js
@@ -1,0 +1,12 @@
+'use strict';
+
+const NodeExternalApi = require('./NodeExternalApi');
+const createDefaultLogger = require('./createDefaultLogger');
+
+const externalApi = new NodeExternalApi();
+
+const defaultLogger = createDefaultLogger({
+  dependencies: { externalApi },
+});
+
+module.exports = defaultLogger;