+++ src/node/transports/ipc.js
@@ -1,12 +1,16 @@
 'use strict';
 
-const electronApi = require('../electronApi');
 const { maxDepth, toJSON } = require('../transforms/object');
 const { transform } = require('../transforms/transform');
 
 module.exports = ipcTransportFactory;
 
-function ipcTransportFactory(logger) {
+/**
+ * @param logger
+ * @param {ElectronExternalApi} externalApi
+ * @returns {transport|null}
+ */
+function ipcTransportFactory(logger, { externalApi }) {
   Object.assign(transport, {
     depth: 3,
     eventId: '__ELECTRON_LOG_IPC__',
@@ -14,10 +18,10 @@ function ipcTransportFactory(logger) {
     transforms: [toJSON, maxDepth],
   });
 
-  return electronApi.isElectron() ? transport : null;
+  return externalApi?.isElectron() ? transport : undefined;
 
   function transport(message) {
-    electronApi.sendIpc(transport.eventId, {
+    externalApi?.sendIpc(transport.eventId, {
       ...message,
       data: transform({ logger, message, transport }),
     });